# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: FeeDLooP API
  description: |
    FeeDLooP is a feedback collection service for software development companies. It provides APIs for managing projects, collecting feedback through embeddable widgets, and managing bug reports, initiatives, and feedback submissions.

    ## Authentication
    Most endpoints require authentication using NextAuth.js session cookies. The widget submission endpoint is public to allow embedding.

    ## Rate Limiting
    API endpoints implement rate limiting to prevent abuse. Refer to security documentation for specific limits.

    ## CORS
    Different CORS policies are applied based on endpoint:
    - Widget endpoints: Permissive CORS for embedding
    - Dashboard/API endpoints: Restrictive CORS for security

    ## Widget Integration
    The embeddable widget uses the `/api/widget/submit` endpoint to submit feedback from any website. See the Framework Integration Guide for implementation details.

  version: 1.0.0
  contact:
    name: FeeDLooP API Support
    url: https://github.com/your-org/feedloop
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:3000
    description: Development server
  - url: https://localhost:3001
    description: Alternative development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/projects/{id}/invitations:
    get:
      tags:
        - Projects
      summary: List project invitations
      description: Returns list of team members and pending invitations for the project
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      responses:
        '200':
          description: List of invitations and team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamMember'
                  pending_invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/PendingInvitation'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (owner only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Projects
      summary: Invite user to project
      description: Sends an invitation to a user to join the project team
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                  format: email
                  description: Email address of user to invite
                role:
                  type: string
                  enum: [member, admin]
                  description: Role to assign to the user
                  default: member
      responses:
        '201':
          description: Invitation sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingInvitation'
        '400':
          description: Validation failed or user already invited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (owner only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # File Upload
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    TeamMember:
      type: object
      required:
        - user_id
        - email
        - role
        - joined_at
      properties:
        user_id:
          type: string
          format: uuid
          description: User identifier
        email:
          type: string
          format: email
          description: User email
        name:
          type: string
          nullable: true
          description: User full name
        role:
          type: string
          enum: [owner, admin, member]
          description: Role in project
        joined_at:
          type: string
          format: date-time
          description: Date joined project
    PendingInvitation:
      type: object
      required:
        - id
        - email
        - role
        - created_at
        - expires_at
      properties:
        id:
          type: string
          format: uuid
          description: Invitation identifier
        email:
          type: string
          format: email
          description: Invited email address
        role:
          type: string
          enum: [admin, member]
          description: Proposed role
        created_at:
          type: string
          format: date-time
          description: Invitation sent timestamp
        expires_at:
          type: string
          format: date-time
          description: Invitation expiry
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Specific error message
          description: Detailed validation errors
        code:
          type: string
          description: Error code for programmatic handling
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
security:
  - sessionAuth: []
...